Update build-aspire-apps-with-python.md
Updated the two C# project references to have one level of indirection to keep with the `hello-python` working directory context of the tutorial, to stay consistent with the first use of that strategy (vs `cd ..` to the parent first). As-is, the two subsequent `dotnet run` commands cannot find the app host project, given the pwd.