Fix for silo not marked dead on graceful shutdown
The silo lifecycle was not tied to the host lifecycle, resulting in silo not marked dead when e.g. deploying a new revision to Azure Container Apps, which in turn results in startup errors in the new revision. Host.UseOrleans() is consistently used in other Orleans examples as a best practice.